moved AudioPlayer to util package and updated effected classes
