+++ src/main/java/com/example/demo/Util/AudioPlayer.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.Util;
 
 import javafx.scene.media.AudioClip;
